requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.1
djongo==1.3.3
idna==2.10
importlib-metadata==2.0.0
isort==5.6.2
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pymongo==3.11.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.3.0
